Re-introduce operation chain length property.
After closer inspection, having the boundary test loop iterate merely over the set of items in its source data entails the significant drawback that a remove for key j+1 will never succeed because the most recent insert could have been only for key j. To fix this would require messing with the index variable. While possible, this would leave the larger drawback of using source data items to guide the inner boundary loop, which is that one inner runner would never accomplish the "up and down" in the number of items within the target map that I want for this feature, unaddressed.

Using the concept of an operation will require some work for synchronizing local and target map state in order to implement an algorithm that efficiently chooses with elements to write, read, or remove, but in order not to add inefficient operations after an application crash, this would have been necessary anyway.