Make cleaner return number of cleaned data structures, and remove comment indicating evaluation of whether a cleaner has been enabled should be made by caller.
Yesterday, I thought the caller should evaluate whether the cleaner has been enabled, and only invoke it, if so. But then again, the cleaner configuration is a piece of knowledge or state only meant for the cleaner itself, and it is the cleaner that has to have the corresponding knowledge of what to do about that configuration knowledge, rather than the caller. Introducing knowledge for operating on this configuration knowledge to the caller would introduce ugly coupling between the caller and the cleaner.