Keep track of set of elements available for insertion as complement to elements currently stored in Hazelcast map.
The boundary test loop's operations for insert, read, and remove actions had complexity O(n) due to necessity to find matching element in source data given only the key of the element. Now, on top of keeping track, for each map, which elements have been inserted, the boundary test loop also keeps track of which elements are still available for insertion. Thus, the complexity of figuring out which map element to use for an insert, read, or remove operation is O(1).