Template configuration blocks individually.
By enabling individual templating of memory blocks, it's possible to decide, for each configuration block that necessitates it, whether the block should be rendered. Currently, this is only necessary for the block referring to native memory -- by enabling users of this chart to easily enable and disable it, they can switch between the community and enterprise edition of Hazelcast without much hassle.