In local cache reflecting state of items corresponding to this test loop in target Hazelcast cluster, store only the element ID rather than the full map key associated with the load element.
Given the element ID, map name, and map number, the full key for storage in as well as retrieval and deletion from the target Hazelcast map can be deterministically calculated. Because all of these data elements are known as soon as an operation on the target Hazelcast map for a given element is to be executed, there's no point storing the full key --> Store only element ID to save memory.

The load element to be acted upon in scope of the action to be executed on the target Hazelcast map, on the other hand, doesn't need to be kept in memory, either, because (a) the key can be calculated, and (b) the payload is now initialized previously for fixed-size mode, too, so there is no case left in which keeping the payload around provides merit. --> Don't store payload, either.