Use slice of bytes to represent random payload rather than string pointer.
This step avoids dereferencing the pointer on each goroutine that wishes to write the underlying string to Hazelcast. Dereferencing was necessary because string pointers get automatically registered with the Hazelcast cluster by means of registration through the Hazelcast client as of version 1.4.2, so they cannot be registered again during later program execution.